refactor: Split Tasks.tsx into smaller components
Priority 1: Tasks.tsx Refactor
- Split 1976-line Tasks.tsx into manageable components:
  - FilterBar: Search/filter controls for task history
  - TaskCard: Individual active task display
  - TaskTable: History table view
  - Pagination: Pagination controls
  - SiteConfigForm: Site configuration form (extracted)
  - ScheduledTasksSection: Scheduled tasks management (extracted)
  - WebCrawlForm, AdhocUrlForm, FileImportForm: Task forms (extracted)
  - FolderBrowser: Directory browser component
  - WebSearchForm, CatalogForm, MarkdownForm, ChunkForm: Task forms (extracted)
- Created FormFields.tsx with shared form components
- Created Tasks.types.ts with shared TypeScript interfaces
- Main Tasks.tsx is now ~400 lines (reduced from 1976)

Priority 2: Loading Skeleton
- Added LoadingSkeleton.tsx with variants: text, card, avatar, button, line
- Includes TaskListSkeleton, TableSkeleton, FormSkeleton helpers

Priority 3: Rate Limiting Middleware
- Added ai_actuarial/api/middleware/rate_limit.py
- Role-based rate limits: guest=10/min, registered=30/min, premium=60/min, operator=200/min
- Applied to search, chat, and collections endpoints
- Includes X-RateLimit-* response headers
- Integrated into FastAPI app middleware stack

Build: npm run build passes successfully
